// resources/js/api/generated/models/getApiV1AdminOperationLogs200DataItem.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiV1AdminOperationLogs200DataItemDeletedAt } from './getApiV1AdminOperationLogs200DataItemDeletedAt';
import type { GetApiV1AdminOperationLogs200DataItemServer } from './getApiV1AdminOperationLogs200DataItemServer';
import type { GetApiV1AdminOperationLogs200DataItemUser } from './getApiV1AdminOperationLogs200DataItemUser';

export type GetApiV1AdminOperationLogs200DataItem = {
  command: string;
  created_at: string;
  deleted_at: GetApiV1AdminOperationLogs200DataItemDeletedAt;
  duration_ms: number;
  end_time: string;
  exit_code: number;
  formatted_date: string;
  /** @minimum 0 */
  id: number;
  is_incomplete: boolean;
  last_message_at: string;
  message_count: number;
  operation_id: string;
  options: string;
  partial_duration_ms: number;
  queued_at: string;
  server: GetApiV1AdminOperationLogs200DataItemServer;
  /** @minimum 0 */
  server_id: number;
  server_name: string;
  services: string;
  stack_name: string;
  start_time: string;
  status: string;
  success: boolean;
  summary: string;
  trigger_source: string;
  updated_at: string;
  user: GetApiV1AdminOperationLogs200DataItemUser;
  /** @minimum 0 */
  user_id: number;
  user_name: string;
};