// resources/js/api/generated/models/getApiV1AdminOperationLogs200Pagination.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV1AdminOperationLogs200Pagination = {
  current_page: number;
  has_next: boolean;
  has_prev: boolean;
  page_size: number;
  total: number;
  total_pages: number;
};