// resources/js/api/generated/models/getApiV1AdminOperationLogsId200.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiV1AdminOperationLogsId200Log } from './getApiV1AdminOperationLogsId200Log';
import type { GetApiV1AdminOperationLogsId200MessagesItem } from './getApiV1AdminOperationLogsId200MessagesItem';

export type GetApiV1AdminOperationLogsId200 = {
  log: GetApiV1AdminOperationLogsId200Log;
  messages: GetApiV1AdminOperationLogsId200MessagesItem[];
};